dead code cleanup, copyright in inits, black on queries, update bsql conda ucx package, remove q28 unused imports

+++ tpcx_bb/xbb_tools/text.py
@@ -20,7 +20,9 @@ import os
 EOL_CHAR = "è"
 
 
-def create_sentences_from_reviews(df, review_column='pr_review_content', end_of_line_char=EOL_CHAR):
+def create_sentences_from_reviews(
+    df, review_column="pr_review_content", end_of_line_char=EOL_CHAR
+):
     import cudf
 
     sentences = df[review_column].str.tokenize(delimiter=end_of_line_char)
@@ -31,34 +33,33 @@ def create_sentences_from_reviews(df, review_column='pr_review_content', end_of_
     # use pr_review_sk as the global position
     # (leaving hardcoded as it's consistent across all queries)
     global_pos = df.pr_review_sk.repeat(tk_cnts).reset_index(drop=True)
-    out = cudf.DataFrame({
-        "sentence": sentences,
-        "review_idx_global_pos": global_pos
-    })
-    out['review_idx_global_pos'] = out['review_idx_global_pos'].astype('int32')
+    out = cudf.DataFrame({"sentence": sentences, "review_idx_global_pos": global_pos})
+    out["review_idx_global_pos"] = out["review_idx_global_pos"].astype("int32")
     return out
 
 
-def create_words_from_sentences(df, sentence_column='sentence', global_position_column='sentence_tokenized_global_pos', delimiter=' '):
+def create_words_from_sentences(
+    df,
+    sentence_column="sentence",
+    global_position_column="sentence_tokenized_global_pos",
+    delimiter=" ",
+):
     import cudf
 
     cleaned_sentences = df[sentence_column].str.replace(
-        [',', ';', '-', '"', '.'],
-        ['', '', '', '', ' '],
-        regex=False)
+        [",", ";", "-", '"', "."], ["", "", "", "", " "], regex=False
+    )
     normalized_sentences = cleaned_sentences.str.normalize_spaces()
     repeat_counts_per_sentence = normalized_sentences.str.token_count(
-        delimiter=delimiter)
+        delimiter=delimiter
+    )
     words = normalized_sentences.str.tokenize(delimiter=delimiter)
 
     # reassociate with the global position
-    global_pos = (df
-                  [global_position_column]
-                  .repeat(repeat_counts_per_sentence)
-                  .reset_index(drop=True)
-                  )
-    out = cudf.DataFrame({
-        "word": words,
-        "sentence_idx_global_pos": global_pos
-    })
+    global_pos = (
+        df[global_position_column]
+        .repeat(repeat_counts_per_sentence)
+        .reset_index(drop=True)
+    )
+    out = cudf.DataFrame({"word": words, "sentence_idx_global_pos": global_pos})
     return out
